Fixing circuit current returning null
Seems like Easee API returns null at some point for the circuitTotalPhaseConductorCurrentL1/L2/L3 which leads to exceptions due to the max() and round_2_dec() functions used in our setup. Didn't get this in the start, but appears repeatedly in my logs now. The null value successfully becomes None, so implemented a check to set 0.0 if state is None.